+++ pages/dashboard/[boardid]/edit.tsx
@@ -1,5 +1,4 @@
 import BasicLayout from '@/src/components/layout/BasicLayout';
-import InvitedDashboardTable from '@/src/components/ui/Table/InvitedDashboardTable';
 import InvitelistTable from '@/src/components/ui/Table/InvitelistTable';
 import MemberTable from '@/src/components/ui/Table/MemberTable';
 
@@ -9,7 +8,6 @@ export default function DashboardEditPage() {
 			<div className='flex flex-col gap-6 p-1'>
 				<MemberTable />
 				<InvitelistTable />
-				<InvitedDashboardTable />
 			</div>
 		</BasicLayout>
 	);
