+++ src/components/dashboard/Card.tsx
@@ -11,30 +11,30 @@ interface CardDataProps {
 
 export default function Card({ cardData }: CardDataProps) {
 	const { title, imageUrl = '', tags, dueDate, assignee } = cardData;
-
+	console.log(cardData);
 	return (
 		<div className='flex w-full flex-col gap-4 '>
-			<div className='flex flex-col justify-between gap-[12px] rounded-[6px] border-[1px] border-solid border-gray3 p-5 sm:p-3 md:flex-row'>
+			<div className='flex flex-col justify-between gap-[12px] rounded-[6px] border-[1px] border-solid border-gray3 bg-white p-5 sm:p-3 md:flex-row'>
 				{imageUrl && (
 					<Image
 						src={imageUrl}
 						width={274}
 						height={160}
-						className='sm:h-[151px] sm:w-[260px] md:h-[53px] md:w-[90px]'
+						className='rounded-md sm:h-[151px] sm:w-[260px] md:h-[53px] md:w-[90px]'
 						alt='사진 예시'
 					/>
 				)}
 				<div
-					className={`flex flex-col gap-[6px] ${imageUrl ? 'md:w-[393px]' : 'md:w-full'}  md:items-start md:justify-center`}
+					className={`flex flex-col gap-[6px] ${imageUrl ? 'md:w-[400px]' : 'sm:w-[257px] md:w-[501px] lg:w-[271px]'}  md:items-start md:justify-center`}
 				>
-					<div className='h-[21px] text-base font-medium leading-5 text-black2 sm:text-[14px]'>
+					<div className='h-[21px] text-left text-base font-medium leading-5 text-black2 sm:text-[14px]'>
 						{title}
 					</div>
 					<div className='md:flex md:w-full'>
 						<div className='flex gap-2 md:flex-row md:gap-[10px]'>
 							{tags.map((tag, index) => (
 								<div key={index} className={`${index >= 2 && 'hidden'}`}>
-									<ColorTagChip>{tags}</ColorTagChip>
+									<ColorTagChip>{tag}</ColorTagChip>
 								</div>
 							))}
 						</div>
@@ -54,8 +54,8 @@ export default function Card({ cardData }: CardDataProps) {
 							</div>
 
 							<DefaultProfileImage
+								classNames='size-6 rounded-full bg-green px-[8px] text-[12px]'
 								nickname={assignee.nickname}
-								classNames='size-6 rounded-full bg-green px-[8px] text-[12px] '
 							/>
 						</div>
 					</div>
